% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpcrcalc.r
\name{gpcrcalc}
\alias{gpcrcalc}
\title{gpcrcalc
Calculates GPCR family enrichment on the input human gene list}
\usage{
custom<-c("TAS1R1","TAS2R3","TAS2R4","PTGDR","PTGDR2","PTGER1","PTGER2","P2RY4","P2RY6","P2RY11","P2RY12","P2RY13","P2RY14","CXCR4","CXCR6","CXCL11","CXCR2","OPN1LW","OPN1MW","OPN1SW","RHO","OPN3","OPN4","OPN5")

res<-gpcrcalc(custom)
}
\arguments{
\item{custom}{a vector of human gene symbol containing possible GPCRs}
}
\description{
gpcrcalc
Calculates GPCR family enrichment on the input human gene list
}
\examples{
custom<-c("TAS1R1","TAS2R3","TAS2R4","PTGDR","PTGDR2","PTGER1","PTGER2","P2RY4","P2RY6","P2RY11","P2RY12","P2RY13","P2RY14","CXCR4","CXCR6","CXCL11","CXCR2","OPN1LW","OPN1MW","OPN1SW","RHO","OPN3","OPN4","OPN5")
res<-gpcrcalc(custom)
res
}
