% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpcrnet.R
\name{gpcrnet}
\alias{gpcrnet}
\title{gpcrnet
built the network of GPCR enriched}
\usage{
custom<-c("TAS1R1","TAS2R3","TAS2R4","PTGDR","PTGDR2","PTGER1","PTGER2","P2RY4","P2RY6","P2RY11","P2RY12","P2RY13","P2RY14","CXCR4","CXCR6","CXCL11","CXCR2","OPN1LW","OPN1MW","OPN1SW","RHO","OPN3","OPN4","OPN5")

res<-gpcrcalc(custom)

gpcrnet(custom,res,layout=layout_as_star,cex=1,distance=2)
}
\arguments{
\item{custom}{a vector of human gene symbol containing possible GPCRs}

\item{res}{a dataframe of results obtained with gpcrcalc function}

\item{layout}{string of layout type for the network (layout_as_star, layout_components, layout_in_circle, layout_nicely,layout_on_grid, layout_on_sphere, layout_randomly, layout_with_dh, layout_with_drl, layout_with_fr, layout_with_gem, layout_with_graphopt, layout_with_kk, layout_with_lgl, layout_with_mds)}

\item{cex}{size of the vertex labels}

\item{distance}{distance between vertex and its label}
}
\description{
gpcrnet
built the network of GPCR enriched
}
\examples{
custom<-c("TAS1R1","TAS2R3","TAS2R4","PTGDR","PTGDR2","PTGER1","PTGER2","P2RY4","P2RY6","P2RY11","P2RY12","P2RY13","P2RY14","CXCR4","CXCR6","CXCL11","CXCR2","OPN1LW","OPN1MW","OPN1SW","RHO","OPN3","OPN4","OPN5")
res<-gpcrcalc(custom)
gpcrnet(custom,res,layout=layout_as_star,cex=1,distance=2)
}
