% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotes.R
\name{plotes}
\alias{plotes}
\title{plotes
Draw the barplot of enrichment score for GPCR families}
\usage{
custom<-c("TAS1R1","TAS2R3","TAS2R4","PTGDR","PTGDR2","PTGER1","PTGER2","P2RY4","P2RY6","P2RY11","P2RY12","P2RY13","P2RY14","CXCR4","CXCR6","CXCL11","CXCR2","OPN1LW","OPN1MW","OPN1SW","RHO","OPN3","OPN4","OPN5")

res<-gpcrcalc(custom)

plotes(res)
}
\arguments{
\item{df}{a dataframe of results obtained with gpcrcalc function}
}
\description{
plotes
Draw the barplot of enrichment score for GPCR families
}
\examples{
custom<-c("TAS1R1","TAS2R3","TAS2R4","PTGDR","PTGDR2","PTGER1","PTGER2","P2RY4","P2RY6","P2RY11","P2RY12","P2RY13","P2RY14","CXCR4","CXCR6","CXCL11","CXCR2","OPN1LW","OPN1MW","OPN1SW","RHO","OPN3","OPN4","OPN5")
res<-gpcrcalc(custom)
plotes(res)
}
